Support Call<T> from Retrofit2
