+++ asyncawait-android/build.gradle
@@ -12,9 +12,11 @@ android {
 
 dependencies {
     compile kotlinStdLib
+    provided retrofit2
 
     testCompile junit
     testCompile expectKt
+    testCompile mockitoKotlin
 
     androidTestCompile junit
     androidTestCompile expectKt
